+++ index.html
@@ -92,7 +92,7 @@
     </div>
 <!-- partial -->
   <script src='https://cdn.jsdelivr.net/npm/uikit@3.21.5/dist/js/uikit.min.js'></script>
-<script src='https://cdn.jsdelivr.net/npm/uikit@3.21.5/dist/js/uikit-icons.min.js'></script><script type="module" src="./script.js"></script>
+<script src='https://cdn.jsdelivr.net/npm/uikit@3.21.5/dist/js/uikit-icons.min.js'></script><script type="module" src="./js/main.js"></script>
 
 </body>
 </html>
